Backend/requirements.txt
Flask
pymongo
stripe
PyJWT
requests
